Module solana_local_cluster::cluster_tests

Structs

GossipVoter

Functions

check_for_new_roots

check_no_new_roots

get_and_verify_slot_entries 🔒

get_client_facing_addr

kill_entry_and_spend_and_verify_rest

poll_all_nodes_for_signature 🔒

send_many_transactions

sleep_n_epochs

spend_and_verify_all_nodes

Spend and verify from every node in the network

start_gossip_voter

Reads votes from gossip and runs them through vote_filter to filter votes that then get passed to generate_vote_tx to create votes that are then pushed into gossip as if sent by a node with identity node_keypair.

submit_vote_to_cluster_gossip

verify_balances

verify_ledger_ticks

verify_slot_ticks 🔒