Struct solana_local_cluster::local_cluster::ClusterConfig

pub struct ClusterConfig {

    pub validator_configs: Vec<ValidatorConfig>,
    pub num_listeners: u64,
    pub validator_keys: Option<Vec<(Arc<Keypair>, bool)>>,
    pub node_stakes: Vec<u64>,
    pub node_vote_keys: Option<Vec<Arc<Keypair>>>,
    pub cluster_lamports: u64,
    pub ticks_per_slot: u64,
    pub slots_per_epoch: u64,
    pub stakers_slot_offset: u64,
    pub skip_warmup_slots: bool,
    pub native_instruction_processors: Vec<(String, Pubkey)>,
    pub cluster_type: ClusterType,
    pub poh_config: PohConfig,
    pub additional_accounts: Vec<(Pubkey, AccountSharedData)>,
    pub tpu_use_quic: bool,
    pub tpu_connection_pool_size: usize,

}

Fields
validator_configs: Vec<ValidatorConfig>

The validator config that should be applied to every node in the cluster

num_listeners: u64

Number of nodes that are unstaked and not voting (a.k.a listening)

validator_keys: Option<Vec<(Arc<Keypair>, bool)>>

List of tuples (pubkeys, in_genesis) of each node if specified. If in_genesis == true, the validator’s vote and stake accounts

node_stakes: Vec<u64>

The stakes of each node

node_vote_keys: Option<Vec<Arc<Keypair>>>

Optional vote keypairs to use for each node

cluster_lamports: u64

The total lamports available to the cluster

ticks_per_slot: u64slots_per_epoch: u64stakers_slot_offset: u64skip_warmup_slots: boolnative_instruction_processors: Vec<(String, Pubkey)>cluster_type: ClusterTypepoh_config: PohConfigadditional_accounts: Vec<(Pubkey, AccountSharedData)>tpu_use_quic: booltpu_connection_pool_size: usize