Struct solana_local_cluster::local_cluster::LocalCluster

pub struct LocalCluster {
    pub funding_keypair: Keypair,
    pub entry_point_info: ContactInfo,
    pub validators: HashMap<Pubkey, ClusterValidatorInfo>,
    pub genesis_config: GenesisConfig,
    pub connection_cache: Arc<ConnectionCache>,
}

Fields
funding_keypair: Keypair

Keypair with funding to participate in the network

entry_point_info: ContactInfo

Entry point from which the rest of the network can be discovered

validators: HashMap<Pubkey, ClusterValidatorInfo>genesis_config: GenesisConfigconnection_cache: Arc<ConnectionCache>
Implementations

impl LocalCluster

pub fn new_with_equal_stakes( num_nodes: usize, cluster_lamports: u64, lamports_per_node: u64, socket_addr_space: SocketAddrSpace ) -> Self

fn sync_ledger_path_across_nested_config_fields( config: &mut ValidatorConfig, ledger_path: &Path )

pub fn new( config: &mut ClusterConfig, socket_addr_space: SocketAddrSpace ) -> Self

pub fn exit(&mut self)

pub fn close_preserve_ledgers(&mut self)

pub fn add_validator_listener( &mut self, validator_config: &ValidatorConfig, stake: u64, validator_keypair: Arc<Keypair>, voting_keypair: Option<Arc<Keypair>>, socket_addr_space: SocketAddrSpace ) -> Pubkey

Set up validator without voting or staking accounts

pub fn add_validator( &mut self, validator_config: &ValidatorConfig, stake: u64, validator_keypair: Arc<Keypair>, voting_keypair: Option<Arc<Keypair>>, socket_addr_space: SocketAddrSpace ) -> Pubkey

Set up validator with voting and staking accounts

fn do_add_validator( &mut self, validator_config: &ValidatorConfig, is_listener: bool, stake: u64, validator_keypair: Arc<Keypair>, voting_keypair: Option<Arc<Keypair>>, socket_addr_space: SocketAddrSpace ) -> Pubkey

pub fn ledger_path(&self, validator_pubkey: &Pubkey) -> PathBuf

fn close(&mut self)

pub fn transfer( &self, source_keypair: &Keypair, dest_pubkey: &Pubkey, lamports: u64 ) -> u64

pub fn check_for_new_roots( &self, num_new_roots: usize, test_name: &str, socket_addr_space: SocketAddrSpace )

pub fn check_no_new_roots( &self, num_slots_to_wait: usize, test_name: &str, socket_addr_space: SocketAddrSpace )

fn transfer_with_client( client: &ThinClient, source_keypair: &Keypair, dest_pubkey: &Pubkey, lamports: u64 ) -> u64

fn setup_vote_and_stake_accounts( client: &ThinClient, vote_account: &Keypair, from_account: &Arc<Keypair>, amount: u64 ) -> Result<()>

pub fn create_dummy_load_only_snapshot_config() -> SnapshotConfig

impl LocalCluster

pub fn wait_for_next_full_snapshot<T>( &self, full_snapshot_archives_dir: T, max_wait_duration: Option<Duration> ) -> FullSnapshotArchiveInfowhere T: AsRef<Path>,

Return the next full snapshot archive info after the cluster’s last processed slot

pub fn wait_for_next_incremental_snapshot( &self, full_snapshot_archives_dir: impl AsRef<Path>, incremental_snapshot_archives_dir: impl AsRef<Path>, max_wait_duration: Option<Duration> ) -> (IncrementalSnapshotArchiveInfo, FullSnapshotArchiveInfo)

Return the next incremental snapshot archive info (and associated full snapshot archive info) after the cluster’s last processed slot

fn wait_for_next_snapshot( &self, full_snapshot_archives_dir: impl AsRef<Path>, incremental_snapshot_archives_dir: Option<impl AsRef<Path>>, next_snapshot_type: NextSnapshotType, max_wait_duration: Option<Duration> ) -> NextSnapshotResult

Return the next snapshot archive infos after the cluster’s last processed slot